+++ pyqgisserver/monitor.py
@@ -19,7 +19,7 @@ class Monitor:
         if name:
             from pyqgisservercontrib.core import componentmanager as cm
             try:
-                service = cm.load_entrypoint('py_qgis_server.monitors',name)()
+                service = cm.load_entrypoint('py_qgis_server.monitors',name).initialize()
                 setattr(cls,'_instance', service)
                 LOGGER.info("Using '%s' monitor service", name)
                 return service